requirements.txt
# Data processing
pandas>=2.2.0
duckdb>=0.10.0
dbt-core>=1.7.0
dbt-duckdb>=1.7.0

# NLP
vaderSentiment>=3.3.2
pysrt>=1.1.2

# Graph analysis
networkx>=3.2

# AI/RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.22
openai>=1.12.0
tiktoken>=0.5.2

# Streamlit
streamlit>=1.31.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Development tools
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
